Ported ThemesTest from laravel 5.3 to Dusk
